requirement.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-classic
sentence-transformers
faiss-cpu
python-dotenv
streamlit
